requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
openai==1.3.7
httpx==0.24.1
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
Werkzeug==2.3.7